Add function without args

+++ Tree.cpp
@@ -22,17 +22,17 @@ Tree::Tree() {
     this->children.resize(0);
 }
 
-void printTree(Tree tree, int num, std::set<int>& se) {
+void printTree(Tree tree, int num, std::set<int> &se) {
     for (int i = 0; i < num; i++) {
         if (se.find(i) != se.end()) std::cout << "|";
         else std::cout << " ";
     }
     std::cout << "|";
     std::cout << tree.node << "\n";
-    se.insert(num+5);
-    for (int j = 0; j < tree.children.size(); j++) {
-        if (j+1 == tree.children.size()) se.erase(num+5);
-        printTree(tree.children[j], num+5, se);
+    se.insert(num + 5);
+    for (size_t j = 0; j < tree.children.size(); j++) {
+        if (j + 1 == tree.children.size()) se.erase(num + 5);
+        printTree(tree.children[j], num + 5, se);
     }
-    se.erase(num+5);
+    se.erase(num + 5);
 }
